Add single large BioSimSpace ABFE node
While this is much less modular than the original approach,
it is much more robust and in keeping with the design of nodes already
in maize-contrib. The new approach takes a list of isomers and runs all
the calculations in parallel. It also avoids loading biosimspace systems
into memory by instead writing biosimspace scripts to be run through
slurm.